Login with PAT command, expired token validation, handlers, tests #147 (#177)

+++ iggy/src/models/pat.rs
@@ -8,5 +8,5 @@ pub struct RawPersonalAccessToken {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PersonalAccessTokenInfo {
     pub name: String,
-    pub expiry: Option<u32>,
+    pub expiry: Option<u64>,
 }
